use leader election to handle multiple instances
The affinity controller already deals with restarts at any point, but parallel
execution of the update requests on the same volume may lead to losing the
volume information.

An easy way around this issue is to use kubernetes leader election, which
is a best-effort coordination between pods. Only the leader is allowed to
actually perform cluster-state changing operations.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>