fix reconciliation in case of interrupted PV deletion
A reconciliation may beinterrupted during the process of PV deletion. There was
an issue where reconciliation did not reset the reclaim policy in these cases.

The issue was that we did not recognize these situations properly. During
deletion of PVs, we have the following phases:

1. The reclaim policy gets patched to "Retain".
2. The PV is deleted.
3. The finalizers are removed to force deletion.

If step 2 or 3 get interrupted, later re-runs did forget to reset the reclaim
policy.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>