prepare for more reconcilers
In order to support more use cases, split the existing affinity reconciler
into a "operations generator" and a "operations executor". With this, we have
a strict divide between looking at the cluster state, deciding what needs to
be done to the resource, and actually making changes.

This should also allow us to write better tests.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>